+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [1.2.0] 2021-04-05
+
+- Allow to call markdown-table-formatter with a glob parameter to find files to format
+- Allow to call markdown-table-formatter without arguments (**/*.md will be used by default to collect files)
+- Fix logs format
+
 ## [1.1.0] 2021-04-05
 
 - Upgrade markdown-table-prettify to 3.4.0 so it can take in account inline disable comments